Additional improvements
