+++ src/vite/plugin.tsx
@@ -2,14 +2,12 @@ import { Plugin, normalizePath } from "vite";
 import { parse } from "es-module-lexer";
 import { cutArrayToLastN } from "../client/utils/common.js";
 import { DevToolsServerConfig } from "../server/config.js";
-import { checkPath, handleDevToolsViteRequest, processPlugins, } from "./utils.js";
+import {   handleDevToolsViteRequest, processPlugins, } from "./utils.js";
 import { ActionEvent, LoaderEvent } from "../server/event-queue.js"; 
 import { RdtClientConfig } from "../client/context/RDTContext.js"; 
-import chalk from "chalk";
-import path from 'path';
-import fs from 'fs';
-import { OpenSourceData } from './types.js';
-import { exec } from "node:child_process";
+import chalk from "chalk"; 
+import { OpenSourceData } from './types.js'; 
+import { DEFAULT_EDITOR_CONFIG, EditorConfig, handleOpenSource } from "./editor.js";
 
 declare global {
   interface Window {
@@ -17,22 +15,7 @@ declare global {
   }
 }
 
-const routeInfo = new Map<string, { loader: LoaderEvent[]; action: ActionEvent[] }>();
-
-type EditorConfig = {
-  name: string;
-  open(path: string | undefined, lineNumber: string): void;
-}
-
-const DEFAULT_EDITOR_CONFIG: EditorConfig = {
-  name: "VSCode",
-  open: (path: string | undefined, lineNumber: string) => {
-    if(!path){
-      return;
-    }
-    exec(`code -g "${normalizePath(path)}${lineNumber}"`);
-  }
-}
+const routeInfo = new Map<string, { loader: LoaderEvent[]; action: ActionEvent[] }>();  
 
 type RemixViteConfig = {
   client?: Partial<RdtClientConfig>;
@@ -48,11 +31,13 @@ type RemixViteConfig = {
 export const defineRdtConfig = (config: RemixViteConfig) =>  config
  
 export const remixDevTools: (args?:RemixViteConfig) => Plugin[] = (args) => {
-  const editor = args?.editor ?? DEFAULT_EDITOR_CONFIG;
+
   const serverConfig = args?.server || {};
-  const clientConfig = args?.client || {
-    editorName: editor.name,
-  };
+  const clientConfig = {
+    ...args?.client,
+    editorName: args?.editor?.name,
+  } 
+ 
   const include = args?.includeInProd ?? false; 
   const improvedConsole = args?.improvedConsole ?? true;
   const remixDir = args?.remixDir || "./app";
@@ -115,57 +100,15 @@ export const remixDevTools: (args?:RemixViteConfig) => Plugin[] = (args) => {
         });
 
         if (!server.config.isProduction) {
-          const openInEditor: EditorConfig['open'] = (path: string | undefined, lineNum: string) => {
+          const editor = args?.editor ?? DEFAULT_EDITOR_CONFIG;
+          const openInEditor = (path: string | undefined, lineNum: string | undefined) => {
             if(!path){
               return;
             }
-
             editor.open(path, lineNum);
           }
          
-          server.hot.on("open-source", ({ data }: OpenSourceData) => {
-            const { source, line, routeID } = data;
-            const lineNum = line ? `:${line}` : "";
-
-            if (source) { 
-              return openInEditor(source, lineNum);
-            }
-
-            if (routeID) {
-              const routePath = path.join(remixDir, routeID);
-              const checkedPath = checkPath(routePath);
-
-              if (!checkedPath) return;
-              const { type, validPath } = checkedPath;
-
-              const reactExtensions = ["tsx", "jsx"];
-              const allExtensions = ["ts", "js", ...reactExtensions];
-              const isRoot = routeID === "root";
-              const findFileByExtension = (prefix: string, filePaths: string[]) =>{
-                const file = filePaths.find(file => allExtensions.some(ext => file === `${prefix}.${ext}`));
-                return file
-              }
-
-              if (isRoot) {
-                if (!fs.existsSync(remixDir)) return;
-                const filesInRemixPath = fs.readdirSync(remixDir);
-                const rootFile = findFileByExtension("root", filesInRemixPath);
-                rootFile && openInEditor(path.join(remixDir, rootFile), lineNum);
-                return;
-              }
-
-              // If its not the root route, then we find the file or folder in the routes folder
-              // We know that the route ID is in the form of "routes/contact" or "routes/user.profile" when is not root
-              // so the ID already contains the "routes" segment, so we just need to find the file or folder in the routes folder
-              if (type === "directory") {
-                const filesInFolderRoute = fs.readdirSync(validPath);
-                const routeFile = findFileByExtension("route", filesInFolderRoute);
-                routeFile && openInEditor(path.join(remixDir, routeID, routeFile), lineNum);
-                return;
-              }
-              return openInEditor(validPath, lineNum);
-            }
-          });
+          server.hot.on("open-source", (data: OpenSourceData) => handleOpenSource({data, openInEditor, remixDir}));
         }
       },
     },
